refactor(FON-468): extract nomination-files-table feature from shared
