+++ apps/client/src/features/agenda/components/AgendaNominationFile.tsx
@@ -5,8 +5,8 @@ import clsx from 'clsx';
 import './AgendaNominationFile.css';
 
 import { Marked } from '@/components/shared/Marked';
-import { NominationFileOutcomeBadge } from '@/components/shared/nomination-files-table/components/cells/nomination-file-outcome/NominationFileOutcomeBadge';
 import { UserAvatarList } from '@/components/shared/user-avatar';
+import { NominationFileOutcomeBadge } from '@/features/nomination-files-table/components/cells/nomination-file-outcome/NominationFileOutcomeBadge';
 import type { FormationEnum } from '@/types/enums.types';
 import type { FoundDocsNominationFiles } from '@api/types';
