+++ apps/client/src/features/nomination-files-table/components/NominationFilesSaveAffectationsButton.tsx
@@ -1,8 +1,8 @@
 import Button from '@codegouvfr/react-dsfr/Button';
 import React from 'react';
 
-import { useAffectations } from '../contexts/files-affectations.context';
-import { useNominationFilesTable } from '../contexts/files-table.context';
+import { useAffectations } from '../context/files-affectations.context';
+import { useNominationFilesTable } from '../context/files-table.context';
 import { useAlerts } from '@/components/shared/alerts/alerts.context';
 import { HttpException } from '@/utils/http-exception';
 import { useAffectNominationFilesReportersMutation } from '@queries/nomination-sessions.queries';
