+++ apps/client/src/features/nomination-files-table/components/cells/magistrat-details/magistrat-comment/MagistratComment.tsx
