+++ apps/client/src/features/nomination-files-table/components/cells/magistrat-details/member-memo/MemberMemo.tsx
