+++ apps/client/src/features/nomination-files-table/components/cells/nomination-file-outcome/NominationFileDocStatusBadge.tsx
