+++ apps/client/src/features/nomination-files-table/components/cells/nomination-file-outcome/nomination-file-outcome-badge.utils.ts
